Add OOP to EditOrderCommand.java
Separate classes that were embedded in EditOrderCommand.java and remove
unused tests from EditOrderCommandTest.java
